views must be long

+++ src/main/java/com/jrestats/viewmodel/Channel.java
@@ -13,7 +13,7 @@ public class Channel {
     public Channel(Map<String, Object> channelsResponse) {
         Map<String, Object> channelItem = DataUtil.getList("items", channelsResponse).get(0);
         this.subscriberCount = FormatUtil.formatNumber(DataUtil.getInteger("statistics.subscriberCount", channelItem));
-        this.viewCount = FormatUtil.formatNumber(DataUtil.getInteger("statistics.viewCount", channelItem));
+        this.viewCount = FormatUtil.formatNumber(DataUtil.getLong("statistics.viewCount", channelItem));
     }
 
 }
